P38-W3: break ry-workspace → ry-checker dependency
Move package metadata types (NamespaceMetadata, namespace_metadata,
attached_packages, native sentinels) and file classification (PackageFileKind,
package_file_kind) from ry-checker to ry-workspace. Move shared constants
(SERIALIZED_BINDINGS_UNENUMERABLE, FFI_PRIMITIVES) to ry-core.

ry-checker now depends on ry-workspace (correct downward direction) and
re-exports these types for backward compatibility.

Add ecosystem/check-cargo-edges.py CI gate to assert forbidden dependency
edges (ry-config and ry-workspace must not depend on ry-checker).

Dependency graph after this change:
  ry-core ← ry-config, ry-workspace, ry-checker
  ry-config ← ry-checker (no ry-checker dep)
  ry-workspace ← ry-checker (no ry-checker dep)

Gates: 990+ workspace tests pass, clippy clean, rustfmt clean, edge check green.
