+++ crates/ry-checker/src/project.rs
@@ -766,62 +766,8 @@ impl Project {
     }
 }
 
-/// The role an R source file has inside its enclosing package.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PackageFileKind {
-    Library,
-    TestCode,
-    TestFixture,
-    Inst,
-    Other,
-}
-
-/// Classify a path relative to its nearest ancestor containing `DESCRIPTION`.
-/// Testthat only sources runner files at `tests/` root and files with its
-/// executable prefixes directly under `tests/testthat/`; deeper R files are
-/// data consumed by tests, not code executed by the package test runner.
-pub fn package_file_kind(path: &std::path::Path) -> PackageFileKind {
-    let Some(root) = path
-        .parent()
-        .and_then(|parent| parent.ancestors().find(|p| p.join("DESCRIPTION").is_file()))
-    else {
-        return PackageFileKind::Other;
-    };
-    let Ok(relative) = path.strip_prefix(root) else {
-        return PackageFileKind::Other;
-    };
-    let components: Vec<_> = relative
-        .components()
-        .filter_map(|component| component.as_os_str().to_str())
-        .collect();
-    match components.as_slice() {
-        ["R", _] => PackageFileKind::Library,
-        ["inst", ..] => PackageFileKind::Inst,
-        ["tests", file] if is_r_source_name(file) => PackageFileKind::TestCode,
-        ["tests", "testthat", file] if is_r_source_name(file) && is_testthat_code_name(file) => {
-            PackageFileKind::TestCode
-        }
-        ["tests", ..] => PackageFileKind::TestFixture,
-        _ => PackageFileKind::Other,
-    }
-}
-
-fn is_r_source_name(name: &str) -> bool {
-    std::path::Path::new(name)
-        .extension()
-        .and_then(|extension| extension.to_str())
-        .is_some_and(|extension| matches!(extension, "R" | "r" | "S" | "s" | "q"))
-}
-
-fn is_testthat_code_name(name: &str) -> bool {
-    let stem = std::path::Path::new(name)
-        .file_stem()
-        .and_then(|stem| stem.to_str())
-        .unwrap_or(name);
-    ["test", "helper", "setup", "teardown"]
-        .iter()
-        .any(|prefix| stem.starts_with(prefix))
-}
+/// File classification — re-exported from ry-workspace (P38-W3).
+pub use ry_workspace::{PackageFileKind, package_file_kind};
 
 /// Whether a file is directly under a package's `R/` directory.
 #[cfg(test)]
